requirements.txt
autogen-agentchat==0.7.4
openai>=1.0.0
pandas>=2.0.0
langdetect>=1.0.9
googletrans>=4.0.0
python-dotenv>=1.0.0
click>=8.0.0
pydantic>=2.0.0